+++ src/renderer/renderer_resource_material_pass.odin
@@ -36,7 +36,7 @@ MaterialPassRef :: common.Ref(MaterialPassResource)
 //---------------------------------------------------------------------------//
 
 InvalidMaterialPassRef := MaterialPassRef {
-	ref = c.UINT64_MAX,
+	ref = c.UINT32_MAX,
 }
 
 //---------------------------------------------------------------------------//
